Rename package to com.ephraim.kigamba.tasktimer
- Rename package name to com.ephraim.kigamba.tasktimer
- Rename app name to com.ephraim.kigamba.tasktimer
- Rename the source package to com.ephraim.kigamba.tasktimer
