+++ app/src/main/res/layout/main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
 	android:layout_height="match_parent"
 	android:fitsSystemWindows="true"
-	tools:context="com.synappze.stoptimer.MainActivity">
+	tools:context=".MainActivity">
 
 	<android.support.design.widget.AppBarLayout
 		android:layout_width="match_parent"
@@ -36,7 +36,7 @@
 			android:paddingLeft="@dimen/activity_horizontal_margin"
 			android:paddingRight="@dimen/activity_horizontal_margin"
 			android:paddingTop="@dimen/activity_vertical_margin"
-			tools:context="com.synappze.stoptimer.MainActivity"
+			tools:context=".MainActivity"
 			tools:showIn="@layout/main">
 
 			<TabHost
